added error + admin board
