+++ L1/AdminBoard/AdminViewController.swift
@@ -1,0 +1,14 @@
+//
+//  AdminViewController.swift
+//  L1
+//
+//  Created by Aleksandrs Sivickis on 10/09/2021.
+//
+
+import UIKit
+
+class AdminViewController:  UIViewController {
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
